# Optimize RunCommand by removing /bin/sh wrapper when possible