--- a/regen.cc
+++ b/regen.cc
@@ -53,6 +53,7 @@ class StampChecker {
 
   struct ShellResult {
     string shell;
+    string shellflag;
     string cmd;
     string result;
     vector<string> missing_dirs;
@@ -232,6 +233,7 @@ class StampChecker {
       ShellResult* sr = new ShellResult;
       commands_.push_back(sr);
       LOAD_STRING(fp, &sr->shell);
+      LOAD_STRING(fp, &sr->shellflag);
       LOAD_STRING(fp, &sr->cmd);
       LOAD_STRING(fp, &sr->result);
       sr->has_condition = LOAD_INT(fp);
@@ -340,7 +342,7 @@ class StampChecker {
 
     COLLECT_STATS_WITH_SLOW_REPORT("shell time (regen)", sr->cmd.c_str());
     string result;
-    RunCommand(sr->shell, sr->cmd, RedirectStderr::DEV_NULL, &result);
+    RunCommand(sr->shell, sr->shellflag, sr->cmd, RedirectStderr::DEV_NULL, &result);
     FormatForCommandSubstitution(&result);
     if (sr->result != result) {
       if (g_flags.dump_kati_stamp) {