For every `$(shell echo "test: $PATH")` command, when `SHELL` is `/bin/bash`, we essentially run: (each arg wrapped in `[]`)

```
[/bin/sh] [-c] [/bin/bash -c "echo \"test: \$PATH\""]
```

This is redundant, since we can just use `SHELL`, and then we don't need to do an extra level of shell escaping either. This change makes us run this instead:

```
[/bin/bash] [-c] [echo "test: $PATH"]
```

If `SHELL` is more complicated than an absolute path to a binary, then we'll fall back to `/bin/sh`.

Using the benchmark introduced in the last change, this reduces a minimal `RunCommand` execution with a simple `SHELL` from 3.7ms to 1.3ms.

For a more complex benchmark (though less normalized), for an AOSP Android build, this change shrinks the average time spent in `$(shell)` functions from 4.5ms to 3ms.